Request GET /whoami to determine if a user is logged in
The current approach consisted in checking that there was a saved token,
however this ignored the fact that the token might be out of date, or
even not valid at all, resulting in false positives from
`resin.auth.isLoggedIn()`.

The alternative is to trigger an HTTP request to GET /whoami, which if
sent a valid token, it will return 200, and 401 otherwise.