Fix Unauthorized error on authenticate()
Resin Request was modified to automatically attempt to refresh the token
once it's validity was considered "soon to be expired".

When using `resin.auth.authenticate()`, Resin Request attempted to
refresh the token, however there is no token at that point, resulting in
a 401 from `/whoami`.

The solution is simply not refreshing the token when attempting to
authenticate.

Fixes: https://github.com/resin-io/resin-sdk/issues/135