Add integration tests
This PR implements an end to end test suite that hits the real API.
Before this, testing in the SDK was made by mocking endpoints and
requests, which allows us to test (some) logic but not the real
interaction the SDK has with the API. The previous test suite was
completely deleted since it holds no benefits against the new one.

Advantages:

- We can easily catch SDK/API incompatibilities and fix them before they
hit our users. This bit us in the past a few times.

- We can build more confidence in the SDK due to higher testing
coverage.

- We could test things that we couldn't meaningfully test before, like
web urls.

Disadvantages:

- The test suite takes a long time to complete. In my system, with a
6MiB/s internet connection, the suite takes ~18m to complete. Given the
test cases interact with a shared resource (a single account), the tests
cannot be parallelised. Almost all of this overhead comes from HTTP
requests to the API made in the tests, and as before/after hooks.

Notice that some functions are still untested at this point (like 2FA,
logs, restart, etc) that require a physical device to be online and/or
need some kind of interaction we can't mock at the moment.

Since the tests hit the API, this means we need to have testing
accounts. All such information is not hardcoded in the test suite and is
instead read from environment variables. The user needs to export the
following variables to be able to run the test suite himself:

- RESINTEST_EMAIL: The main account email
- RESINTEST_PASSWORD: The main account password
- RESINTEST_USERNAME: The main account username.
- RESINTEST_USERID: The main account user id.

The following environment variables represent a *different* account,
used as a facility to test user registration. Notice that the account
represented by the following variables **will be completely erased as
part of the tests**.

- RESINTEST_REGISTER_EMAIL: The email of the account to register.
- RESINTEST_REGISTER_PASSWORD: The password of the account to register.
- RESINTEST_REGISTER_USERNAME: The username of the account to register.

Failure in providing *any* of the above mentioned variables will result
in an error and the test suite will exit with a non-successful error
code.

Extra:

- Some very trivial changes were introduced in some of the functions
(mainly documention) as a result of the test suite implementation.