Implement resin.models.device.getStatus()
This function makes use of [resin-device-status](https://github.com/resin-io/resin-device-status)
to return a slug representing the current device status, based on the
properties of the device object.

Only the device status slug is returned, instead of an object containing
both the slug and the display name to simplify the interface and return
a unique identifier that the user can then choose to show with the
display name of their preference.

The function, despite being synchronous, was converted into a promise to
be consistent with the rest of the API, and to avoid potential breaking
changes by this function needing to be async in the future.

Fixes: https://github.com/resin-io/resin-sdk/issues/183