Stop reading settings from the filesystem
This is a critical commit towards porting the SDK to the browser.
Currently, the Resin SDK would scan the file system and environment
variables to configure itself with the `resin-settings-client` project,
but this coupling to the environment prevents the SDK to run in an
environment like the browser without major hacks/workarounds.

The plan is that the Resin CLI, the biggest SDK client, will be the one
in charge of reading configuration from the environment, while the SDK
will only maintain some sane defaults and allow the client to override
them, or set custom ones.

In the future, the Resin UI, another big player to depend on the SDK,
will handle configuration in a way it makes sense for a web application.