Explicitly pass baseUrl in resin-request calls
This is the starting point to decouple `resin-request` from SDK
settings, in order to adapt to the new way of managing settings.

Previously, `resin-request` computed the full url from a relative url by
using `resin-settings-client`. Passing a `baseUrl` that's equal to the
API url leads to the exact same behaviour, but now this comes from the
SDK to `resin-request` instead of backwards, improving decoupleness.

We upgrade `resin-request` to v4.0.2 since that version contains a fix
to make `baseUrl` work as expected.

See: https://github.com/resin-io/resin-request/pull/49