Make sure shorter uuids are expanded before passing to pine
If the user passes a shorter uuid, and then we use this shorter uuid as
a pine filter, it will match nothing, causing no errors but no real
action, which can be very hard to debug.

As a fix, we call `resin.models.device.get()` with the shorter uuid, and
access the resulting `.uuid` field, which is ensured to be the full one,
before passing it to pine.

Signed-off-by: Juan Cruz Viotti <jviottidc@gmail.com>