Improve performance by select only the fields required where possible
This reduces load on the API, and traffic on the wire, by transferring
only the data on objects that we need, rather than the full object.

This isn't done for every case yet: this covers all the obvious and
automatically tested cases, and there's definitely more we can do in
future. This also doesn't do this for any cases that would be sensible
but breaking - that's a whole separate problem. Adding now to get
the ball rolling on this though.

Change-Type: patch