tests: Use RPi2 for the device.getDisplayName() tests
At some point we changed the name field of RPI1 and
had to update these tests to use the new one. We
also recently reverted some OS releases, which as a
result bought back the old name and broke the tests
again. So, I decided to just use RPi2 for this tests
since we don't expect its name to change soon but
also its one of the few devices that also has an alias.

Change-type: patch
Signed-off-by: Thodoris Greasidis <thodoris@balena.io>