Allow retrieving applications by application case insensitive slug
This allows users to disambiguate shared
applications that have the same name in a more
user friendly way than using the application id.
The application slug equals
`${owner}/${app_name}`.toLowerCase() on the
cloud api and it's just app_name.toLowerCase() in
open balena, so we can be sure that the $or will
never return more than one items when a slug is
provided.

Resolves: #795
Change-type: minor
Signed-off-by: Thodoris Greasidis <thodoris@balena.io>