Prevent invalid $selects in strict pine.get variant calls
Used the NoInfer helper to prevent TS from using the
old pine.get overloads, unless the generic
parameters are explicitely provided. This way when
no generic arguments are provided TS will try to
match only the fully typed variants and in case of an
error it will output the reason from those overloads.
Previously it would fallback to the old pine.get
variants and try to infer a loose generic type, which
wasn't correct and silenced some errors This is a bit
aggressive change and as a result I decided to limit
this change for the strict pine variant, which is opt-in,
and hold it back from the main typings until the next
major.

Change-type: minor
Signed-off-by: Thodoris Greasidis <thodoris@balena.io>