Bump pinejs-client-js to v8.2.0
Update pinejs-client-js from 7.0.0 to 8.2.0

This is a fairly extensive change for all the SDK typings.
As a general rule, it is now using the incoming types from balena-api rather than manually mantaing its own.

These types are in geral less strict than the previous types, however, are in nature shared between server and client components, and thus any improvement applies to the entire stack.

The bigger changes are:
 - The resource typings (e.g. Application) now are presented as `Application['Read']` & `Application['Write']`
 - The pinejs-client-js typings are all directly exposed inside the `BalenaSDK.Pine.*` scope
 - PineTypedResult should be replaced with Pine.OptionsToResponse interface
Several other small changes apply, specially the ones from pinejs-client-core (most noticeable drop of $filter with the format of {a: b: c} for navigation resources and preferences for object format for $orderby)

Change-type: major